app/requirements.txt
Flask
pymongo
dnspython
python-dotenv

pytest
pytest-flask
mongomock